+++ apps/products-feed/src/modules/editor-js/readme.md
@@ -1,0 +1,3 @@
+Editor.js is used in Dashboard for product description. Feed description should be rendered to standard text.
+
+This module contains logic that transforms Editor.js to plain text
