+++ apps/products-feed/src/modules/google-feed/fetch-product-data.ts
