+++ apps/products-feed/src/modules/google-feed/types.ts
